broke the redux file in smaller pieces so it is more readable, to this end an Authentication and Books folder was created with its corresponding actions and reducers. Created a style.css for the dashboard
