+++ wbooks/src/app/redux/actionTypes.js
@@ -1,0 +1,14 @@
+const actionTypes = {
+  CHANGE_FILTER: 'CHANGE_FILTER',
+  LOGIN_REQUEST : 'LOGIN_REQUEST ',
+  LOGIN_SUCCESS: 'LOGIN_SUCCESS',
+  LOGIN_FAILURE: 'LOGIN_FAILURE',
+  BOOKS_LOADING: 'BOOKS_LOADING',
+  BOOKS_LOADED: 'BOOKS_LOADED',
+  BOOKS_FAILED: 'BOOKS_FAILED',
+  BOOK_LOADING: 'BOOK_LOADING',
+  BOOK_LOADED: 'BOOK_LOADED',
+  BOOK_FAILED: 'BOOK_FAILED',
+  LOGOUT: 'LOGOUT',
+}
+export default actionTypes;
