+++ wbooks/src/app/redux/store.js
@@ -1,6 +1,9 @@
 import thunk from 'redux-thunk';
-import { RootReducer } from './';
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, combineReducers } from 'redux';
+import auth from './Authentication/reducers';
+import books from './Books/reducers';
+
+const RootReducer = combineReducers({auth, books});
 
 const store = createStore(RootReducer, applyMiddleware(thunk));
 export default store;
