+++ wbooks/src/app/screens/dashboard/style.css
@@ -1,0 +1,3 @@
+.no-books-found {
+  margin: 20;
+}
